01. Starting Project
Let's review what we learned in this class:

How to read a layout in Figma and plan how its structure in code will be;
The process of prioritizing content before styling;
What are Widgets content;
Widgets Flutter and Widgets Material.
That was our big kick-off, see you in the next class!