03. Creating new Components
Let's review what we learned in this class:

Perceive visual and structural repetitions in the project, applying techniques to organize the code according to the layout;
Plan the structure of a customized widget to suit the needs of the project;
Receive widgets within other widgets, thinking about making our customized widget as flexible as possible;
Know the shading property of Widgets to add a new visual effect.
Now it's time to learn how to deal with themes, an important concept to further style an application's layout.