04. Themes Implementing
We have completed an important stage! In this class, we learn to:

Understand what themes are and what they are for;
See patterns for creating a customized theme;
Establish the colors that will be used in the project in an isolated file;
Develop a new theme with our colors and typographic properties;
Adapt our project to use the stylizations of our new theme.