05. Implementing the new section
We have taken another step forward in implementation! In this class, we learned to:

Differentiate visual elements from elements with content;
Understand the needs to create specific widgets for these visual elements;
Apply stylized widgets in practice: small colored spheres, line to separate sections of the content;
Practice the concept of using all the resources we have prepared so far to finalize the section of RecentActivity.
Now you've unlocked new skills to style your apps' screen.